+++ qlab/problems/__init__.py
@@ -6,7 +6,8 @@ from qlab.problems import (  # noqa: F401  (import = registration)
     bernstein_vazirani,
     deutsch_jozsa,
     maxcut,
+    simon,
     state_prep,
 )
 
-__all__ = ["state_prep", "maxcut", "bernstein_vazirani", "deutsch_jozsa"]
+__all__ = ["state_prep", "maxcut", "bernstein_vazirani", "deutsch_jozsa", "simon"]